pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["stepan <stepanlezennikov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pre-commit = "^4.0.1"
pytest = "^8.3.4"
django-environ = "^0.11.2"
psycopg = "^3.2.3"
ruff = "^0.8.2"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
celery = "^5.4.0"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
jwt = "^1.3.1"
redis = "^5.2.1"
django-celery-results = "^2.5.1"
requests = "^2.32.3"
mypy = "^1.14.1"
black = "^24.10.0"
django-stubs = "^5.1.1"
types-requests = "^2.32.0.20241016"
types-python-dateutil = "^2.9.0.20241206"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
